requirements.txt
pytz==2012d
cassandra-driver==3.4.0